backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
python-dotenv==1.0.1
scapy==2.5.0
aiosqlite==0.20.0
